requirements.txt
setuptools==68.0.0
PyYAML~=6.0.1
pytest~=7.4.0